requirements.txt
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
decorator==4.4.2
h11==0.14.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
isodate==0.6.1
lxml==5.2.2
mypy==1.10.1
mypy-extensions==1.0.0
numpy==2.0.0
opencv-contrib-python==4.10.0.84
outcome==1.3.0.post0
pillow==10.4.0
proglog==0.1.10
pycountry==24.6.1
pycryptodome==3.20.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==70.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
streamlink==6.8.2
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
